fix(array): 🐛  Resolve issues with circular dependency
